fix runtime hardening gaps and align CI lint/benchmark reporting (#18)
* fix runtime hardening gaps and align CI lint/benchmark reporting

* fix: failed ci

* Added xtask audit

* Fix: xtask ci issues

+++ crates/prkdb-core/src/wal/async_log_segment.rs
@@ -214,7 +214,7 @@ impl AsyncLogSegment {
 
         // Get approximate file position from index
         let index = self.index.lock().await;
-        let file_pos = index.lookup(offset).unwrap_or(0) as u64;
+        let file_pos = index.lookup(offset).unwrap_or(0);
         drop(index);
 
         // Scan from that position
